logica v1 Casos de uso
reestructuracion de canvas para facilitar el guardado de estados e informacion
