+++ src/components/casosDeUso/Figura.js
@@ -1,0 +1,13 @@
+class Figura extends React.Component {
+    constructor(props) {
+      super(props);
+  
+      this.state = { 
+        nombre:"",
+        x:0,
+        y:0,
+
+        
+      };
+    }
+}
